Testing header presensce
Allowing to easily determine if a header is actually present in the HTTP
connection. This is useful to determine if when using `best_of` the fallback
should be a global fallback for the header or whatever is returned by
`best_of`.

This avoids relying on the ordering of options passed to `best_of`, as without a
presence check the selected option would be the first option passed.